pyproject.toml
[project]
name = "tasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
]
